# Improve sample