--- a/src/simulations/sample.jl
+++ b/src/simulations/sample.jl
@@ -74,7 +74,7 @@ function clean_space!(model::ABM)
             model.space.agent_positions[node] = Int[]
         end
         for (k, v) in model.agents
-            push!(model.space.agent_positions[pos_vertex(v.pos, model)], v.id)
+            push!(model.space.agent_positions[coord2vertex(v.pos, model)], v.id)
         end
     end
 end