--- a/test/benchmark/benchmark.jl
+++ b/test/benchmark/benchmark.jl
@@ -25,7 +25,28 @@ include("schelling_defs.jl")
 a = @benchmark data=step!(model, agent_step!, 10, [:pos, :mood, :group], when=1:10) setup=(model = instantiate_modelS(numagents=370, griddims=(20,20), min_to_be_happy=5))
 display(a)
 
-println("\n\nsample!")
+println("\n\nsample! 1 - no space - n == numagents")
 include("sample_defs.jl")
-a = @benchmark sample!(m, nagents(m)) setup=(m = sampleModelInitialize())
+a = @benchmark sample!(m, nagents(m)) setup=(m = sampleModelInitialize(n=100))
 display(a)
+
+println("\n\nsample! 2 - no space - n == numagents*2")
+a = @benchmark sample!(m, nagents(m)*2) setup=(m = sampleModelInitialize(n=100))
+display(a)
+
+println("\n\nsample! 3 - no space - n == numagents/2")
+a = @benchmark sample!(m, 50) setup=(m = sampleModelInitialize(n=100))
+display(a)
+
+println("\n\nsample! 4 - with space - n == numagents")
+a = @benchmark sample!(m, nagents(m)) setup=(m = sampleModelInitializeS(n=100))
+display(a)
+
+println("\n\nsample! 5 - with space - n == numagents * 2")
+a = @benchmark sample!(m, nagents(m)*2) setup=(m = sampleModelInitializeS(n=100))
+display(a)
+
+println("\n\nsample! 6 - with space - n == numagents/2")
+a = @benchmark sample!(m, 50) setup=(m = sampleModelInitializeS(n=100))
+display(a)
+