This PR improves `sample!`'s performance by about 30%. It achieves it by not creating a temporary array of agents.

Here are some benchmarks:

## master branch

```
sample! 1 - no space - n == numagents
BenchmarkTools.Trial: 
  memory estimate:  49.56 KiB
  allocs estimate:  606
  --------------
  minimum time:     174.100 μs (0.00% GC)
  median time:      177.901 μs (0.00% GC)
  mean time:        196.164 μs (1.03% GC)
  maximum time:     2.224 ms (89.64% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 2 - no space - n == numagents*2
BenchmarkTools.Trial: 
  memory estimate:  115.61 KiB
  allocs estimate:  1209
  --------------
  minimum time:     351.200 μs (0.00% GC)
  median time:      363.301 μs (0.00% GC)
  mean time:        401.761 μs (1.50% GC)
  maximum time:     3.361 ms (86.03% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 3 - no space - n == numagents/2
BenchmarkTools.Trial:
  memory estimate:  25.34 KiB
  allocs estimate:  306
  --------------
  minimum time:     88.000 μs (0.00% GC)
  median time:      92.200 μs (0.00% GC)
  mean time:        104.723 μs (1.23% GC)
  maximum time:     3.058 ms (96.21% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 4 - with space - n == numagents
BenchmarkTools.Trial:
  memory estimate:  54.25 KiB
  allocs estimate:  706
  --------------
  minimum time:     169.700 μs (0.00% GC)
  median time:      175.100 μs (0.00% GC)
  mean time:        197.991 μs (1.63% GC)
  maximum time:     2.978 ms (93.36% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 5 - with space - n == numagents * 2
BenchmarkTools.Trial:
  memory estimate:  126.84 KiB
  allocs estimate:  1413
  --------------
  minimum time:     336.400 μs (0.00% GC)
  median time:      360.250 μs (0.00% GC)
  mean time:        401.352 μs (2.20% GC)
  maximum time:     5.884 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 6 - with space - n == numagents/2
BenchmarkTools.Trial:
  memory estimate:  27.69 KiB
  allocs estimate:  356
  --------------
  minimum time:     88.999 μs (0.00% GC)
  median time:      92.001 μs (0.00% GC)
  mean time:        104.519 μs (2.07% GC)
  maximum time:     3.379 ms (95.53% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

## PR

```
sample! 1 - no space - n == numagents
BenchmarkTools.Trial: 
  memory estimate:  42.39 KiB
  allocs estimate:  403
  --------------
  minimum time:     120.400 μs (0.00% GC)
  median time:      142.100 μs (0.00% GC)
  mean time:        166.896 μs (1.22% GC)
  maximum time:     6.078 ms (96.76% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 2 - no space - n == numagents*2
BenchmarkTools.Trial: 
  memory estimate:  101.30 KiB
  allocs estimate:  806
  --------------
  minimum time:     245.800 μs (0.00% GC)
  median time:      257.800 μs (0.00% GC)
  mean time:        288.679 μs (1.94% GC)
  maximum time:     5.541 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 3 - no space - n == numagents/2
BenchmarkTools.Trial: 
  memory estimate:  21.69 KiB
  allocs estimate:  203
  --------------
  minimum time:     61.000 μs (0.00% GC)
  median time:      64.901 μs (0.00% GC)
  mean time:        72.084 μs (1.31% GC)
  maximum time:     2.512 ms (96.25% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 4 - with space - n == numagents
BenchmarkTools.Trial: 
  memory estimate:  52.39 KiB
  allocs estimate:  625
  --------------
  minimum time:     131.400 μs (0.00% GC)
  median time:      136.900 μs (0.00% GC)
  mean time:        156.797 μs (1.17% GC)
  maximum time:     2.605 ms (93.35% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 5 - with space - n == numagents * 2
BenchmarkTools.Trial: 
  memory estimate:  120.84 KiB
  allocs estimate:  1233
  --------------
  minimum time:     268.799 μs (0.00% GC)
  median time:      288.450 μs (0.00% GC)
  mean time:        328.640 μs (2.60% GC)
  maximum time:     3.923 ms (77.71% GC)
  --------------
  samples:          10000
  evals/sample:     1

sample! 6 - with space - n == numagents/2
BenchmarkTools.Trial: 
  memory estimate:  26.86 KiB
  allocs estimate:  320
  --------------
  minimum time:     70.901 μs (0.00% GC)
  median time:      74.399 μs (0.00% GC)
  mean time:        83.728 μs (1.09% GC)
  maximum time:     3.451 ms (94.62% GC)
  --------------
  samples:          10000
  evals/sample:     1
```